fix(telegram): unify durable ingress admission
