+++ extensions/telegram/src/telegram-ingress-spool.ts
@@ -4,9 +4,7 @@ import path from "node:path";
 import type { ChannelIngressQueue } from "openclaw/plugin-sdk/channel-outbound";
 import { computeBackoff, type BackoffPolicy } from "openclaw/plugin-sdk/runtime-env";
 import { resolveStateDir } from "openclaw/plugin-sdk/state-paths";
-import type { TelegramBotInfo } from "./bot-info.js";
 import { getTelegramRuntime } from "./runtime.js";
-import { getTelegramSequentialKey } from "./sequential-key.js";
 import { normalizeTelegramStateAccountId } from "./state-account-id.js";
 import {
   TELEGRAM_SPOOLED_UPDATE_PAYLOAD_VERSION,
@@ -79,45 +77,6 @@ export function openTelegramIngressQueue(
     accountId: parts.accountId,
     stateDir: parts.stateDir,
   });
-}
-
-export function telegramSpooledUpdateLaneKey(update: unknown, botInfo?: TelegramBotInfo): string {
-  return getTelegramSequentialKey({
-    update: update as Parameters<typeof getTelegramSequentialKey>[0]["update"],
-    ...(botInfo ? { me: botInfo } : {}),
-  });
-}
-
-/**
- * Durable-before-ack accept path: commit the update to the ingress queue.
- * Polling advances offset only after this returns; webhook returns 200 only after.
- */
-export async function writeTelegramSpooledUpdate(params: {
-  spoolDir: string;
-  update: unknown;
-  laneKey?: string;
-  now?: number;
-}): Promise<number> {
-  const updateId = resolveTelegramUpdateId(params.update);
-  if (updateId === null) {
-    throw new Error("Telegram update missing numeric update_id.");
-  }
-  const receivedAt = params.now ?? Date.now();
-  const queue = openTelegramIngressQueue(params.spoolDir);
-  await queue.enqueue(
-    telegramQueueEventId(updateId),
-    {
-      version: TELEGRAM_SPOOLED_UPDATE_PAYLOAD_VERSION,
-      updateId,
-      receivedAt,
-      update: params.update,
-    },
-    {
-      receivedAt,
-      laneKey: params.laneKey ?? telegramSpooledUpdateLaneKey(params.update),
-    },
-  );
-  return updateId;
 }
 
 /** Backoff for irrevocable-adoption completion retries (bot-message only). */
